+++ tap_outreach/sync.py
@@ -35,6 +35,7 @@ STREAM_CONFIGS = {
             'callDispositionId',
             'callPurposeId',
             'opportunityId',
+            'phoneNumberId',
             'prospectId',
             'sequenceId',
             'sequenceStateId',
@@ -57,7 +58,15 @@ STREAM_CONFIGS = {
         'url_path': 'events',
         'replication': 'incremental',
         'filter_field': 'eventAt',
-        'fks': ['prospectId', 'userId']
+        'fks': [
+            'accountId',
+            'callId',
+            'mailingId',
+            'opportunityId',
+            'prospectId',
+            'taskId',
+            'userId'
+        ]
     },
     'mailboxes': {
         'url_path': 'mailboxes',
@@ -70,7 +79,7 @@ STREAM_CONFIGS = {
         'replication': 'incremental',
         'filter_field': 'updatedAt',
         'fks': [
-            'calendarId',
+            'followUpSequenceId',
             'mailboxId',
             'opportunityId',
             'prospectId',
@@ -89,7 +98,9 @@ STREAM_CONFIGS = {
             'accountId',
             'creatorId',
             'opportunityStageId',
-            'ownerId'
+            'ownerId',
+            'stageId',
+            'updaterId'
         ]
     },
     'personas': {
@@ -104,40 +115,52 @@ STREAM_CONFIGS = {
         'fks': [
             'accountId',
             'creatorId',
-            'defaultPluginMappingId',
             'ownerId',
             'personaId',
             'stageId',
             'updaterId'
         ]
     },
-    'stages': {
-        'url_path': 'stages',
-        'replication': 'incremental',
-        'filter_field': 'updatedAt',
-        'fks': ['creatorId', 'updaterId']
-    },
-    'sequences': {
-        'url_path': 'sequences',
-        'replication': 'incremental',
-        'filter_field': 'updatedAt',
-        'fks': ['creatorId', 'ownerid', 'updaterId']
-    },
     'sequence_states': {
         'url_path': 'sequenceStates',
         'replication': 'incremental',
         'filter_field': 'updatedAt',
-        'fks': ['accountid', 'creatorId', 'prospectId', 'sequenceId']
+        'fks': [
+            'accountId',
+            'creatorId',
+            'mailboxId',
+            'opportunityId',
+            'prospectId',
+            'sequenceId',
+            'sequenceStepId'
+        ]
     },
     'sequence_steps': {
         'url_path': 'sequenceSteps',
         'replication': 'incremental',
         'filter_field': 'updatedAt',
-        'fks': ['creatorId', 'sequenceId', 'updaterId']
+        'fks': [
+            'callPurposeId',
+            'creatorId',
+            'sequenceId',
+            'taskPriorityId',
+            'updaterId'
+        ]
     },
     'sequence_templates': {
         'url_path': 'sequenceTemplates',
         'replication': 'full',
+        'fks': ['creatorId', 'sequenceStepId', 'templateId', 'updaterId']
+    },
+    'sequences': {
+        'url_path': 'sequences',
+        'replication': 'incremental',
+        'filter_field': 'updatedAt',
+        'fks': ['creatorId', 'ownerId', 'rulesetId', 'updaterId']
+    },
+    'stages': {
+        'url_path': 'stages',
+        'replication': 'incremental',
         'filter_field': 'updatedAt',
         'fks': ['creatorId', 'updaterId']
     },
@@ -157,15 +180,16 @@ STREAM_CONFIGS = {
             'sequenceId',
             'sequenceStateId',
             'sequenceStepId',
+            'sequenceTemplateId',
             'subjectId',
             'taskPriorityId',
-            'taskThemeId',
             'templateId'
         ]
     },
     'teams': {
         'url_path': 'teams',
-        'replication': 'full',
+        'replication': 'incremental',
+        'filter_field': 'updatedAt',
         'fks': ['creatorId', 'updaterId']
     },
     'users': {
@@ -173,7 +197,6 @@ STREAM_CONFIGS = {
         'replication': 'incremental',
         'filter_field': 'updatedAt',
         'fks': [
-            'calendarId',
             'mailboxId',
             'profileId',
             'roleId',
@@ -185,13 +208,13 @@ STREAM_CONFIGS = {
 
 
 def get_bookmark(state, stream_name, default):
-    return state.get('bookmarks', {}).get(stream_name, default)
+    return state.get('bookmarks', {}).get(stream_name, {}).get(STREAM_CONFIGS[stream_name].get('filter_field'), default)
 
 
 def write_bookmark(state, stream_name, value):
     if 'bookmarks' not in state:
         state['bookmarks'] = {}
-    state['bookmarks'][stream_name] = value
+    state['bookmarks'][stream_name] = {STREAM_CONFIGS[stream_name]['filter_field']: value}
     singer.write_state(state)
 
 
@@ -227,14 +250,7 @@ def process_records(stream, mdata, max_modified, records, filter_field, fks):
                             raise Exception(
                                 'null or `id` field expected for `data` relationship')
 
-                        # potential fix for the issue - https://github.com/singer-io/tap-outreach/issues/20
-                        if stream.tap_stream_id != "prospects":
-                            if fk_field_name in record_flat:
-                                raise Exception('`{}` exists as both an attribute and generated relationship name'.format(fk_field_name))
-
-                        if data_value == None:
-                            record_flat[fk_field_name] = None
-                        else:
+                        if data_value is not None:
                             record_flat[fk_field_name] = data_value['id']
 
             if filter_field in record_flat and record_flat[filter_field] > max_modified:
@@ -249,7 +265,7 @@ def process_records(stream, mdata, max_modified, records, filter_field, fks):
         return max_modified
 
 
-def sync_endpoint(client, config, catalog, state, start_date, stream, mdata):
+def sync_endpoint(client, config, state, start_date, stream, mdata):
     stream_name = stream.tap_stream_id
     last_datetime = get_bookmark(state, stream_name, start_date)
 
@@ -281,12 +297,8 @@ def sync_endpoint(client, config, catalog, state, start_date, stream, mdata):
                     filter_field)] = '{}..inf'.format(paginate_datetime)
                 query_params['sort'] = filter_field
 
-        LOGGER.info('{} - Syncing data since {} - page: {}, limit: {}, offset: {}'.format(
-            stream.tap_stream_id,
-            last_datetime,
-            page,
-            count,
-            offset))
+        LOGGER.info(f"{stream.tap_stream_id} - Syncing data since {last_datetime} \
+                    - page: {page}, limit: {count}, offset: {offset}")
 
         querystring = '&'.join(['%s=%s' % (key, value)
                                 for (key, value) in query_params.items()])
@@ -323,18 +335,18 @@ def sync_endpoint(client, config, catalog, state, start_date, stream, mdata):
 
 
 def update_current_stream(state, stream_name=None):
+    LOGGER.info("Setting the stream as currently syncing")
     set_currently_syncing(state, stream_name)
     singer.write_state(state)
 
 
 def sync(client, config, catalog, state, start_date):
     selected_streams = catalog.get_selected_streams(state)
-    selected_streams = sorted(selected_streams, key=lambda x: x.tap_stream_id)
 
     for stream in selected_streams:
         mdata = metadata.to_map(stream.metadata)
         update_current_stream(state, stream.tap_stream_id)
-        sync_endpoint(client, config, catalog, state,
+        sync_endpoint(client, config, state,
                       start_date, stream, mdata)
 
     update_current_stream(state)